refactor: extract parseCikSafely into a shared util
Form_C.storage.ts and Form_D.storage.ts each carried a near-identical
parseCikSafely helper, and pipeline-test-util.ts grew a third copy
(named safeCikToInt) while explicitly noting it mirrored those.

Move the canonical implementation (Form_D's, which handles both string
and number inputs) into src/util/parseCik.ts and import from all three
sites. The test util re-exports it under the existing safeCikToInt name
so callers don't have to change.
