+++ src/sec/forms/exempt-offerings/Form_C.storage.ts
@@ -13,25 +13,11 @@ import { CrowdfundingTemporalRepo } from "../../../storage/portal/CrowdfundingTe
 import type { Crowdfunding, CrowdfundingOfferings, CrowdfundingReports } from "../../../storage/portal/CrowdfundingSchema";
 import { isBadPersonField } from "../../../types/edgar/bad-data";
 import type { FormC } from "./Form_C.schema";
+import { parseCikSafely } from "../../../util/parseCik";
 
 const RELATION_TYPE_ISSUER = "form-c:issuer";
 const RELATION_TYPE_CO_ISSUER = "form-c:co-issuer";
 const RELATION_TYPE_SIGNATURE = "form-c:signature";
-
-/**
- * Coerce a CIK-shaped string into a non-negative integer, or 0 if the input is
- * missing/non-numeric. EDGAR occasionally emits non-digit cruft (whitespace,
- * stray punctuation) that `parseInt` would silently turn into `NaN`.
- */
-function parseCikSafely(raw: string | undefined | null): number {
-  if (!raw) return 0;
-  const trimmed = String(raw).trim();
-  // Require all digits — parseInt would accept "123abc" as 123, which is a
-  // plausible but wrong CIK.
-  if (!/^\d+$/.test(trimmed)) return 0;
-  const parsed = Number(trimmed);
-  return Number.isFinite(parsed) ? parsed : 0;
-}
 
 /**
  * Parse a compensation/financial interest string to extract a percentage and detail text.
