+++ src/sec/forms/exempt-offerings/Form_D.storage.ts
@@ -24,6 +24,7 @@ import {
 } from "./Form_D.schema";
 import { InvestmentOffering } from "../../../storage/investment-offering/InvestmentOfferingSchema";
 import { InvestmentOfferingHistory } from "../../../storage/investment-offering/InvestmentOfferingHistorySchema";
+import { parseCikSafely } from "../../../util/parseCik";
 
 /**
  * Coerce a numeric-shaped string into a finite integer or null. EDGAR-emitted
@@ -37,14 +38,6 @@ function parseIntegerOrNull(raw: string | number | undefined | null): number | n
   if (!/^-?\d+$/.test(trimmed)) return null;
   const parsed = Number(trimmed);
   return Number.isFinite(parsed) ? parsed : null;
-}
-
-function parseCikSafely(raw: string | number | undefined | null): number {
-  if (raw === undefined || raw === null) return 0;
-  const trimmed = String(raw).trim();
-  if (!/^\d+$/.test(trimmed)) return 0;
-  const parsed = Number(trimmed);
-  return Number.isFinite(parsed) ? parsed : 0;
 }
 
 // relation types for form-d
@@ -260,10 +253,9 @@ async function processIssuer(cik: number, issuer: Issuer, isPrimaryIssuer: boole
       company.company_hash_id
     );
   } catch (error) {
+    const message = error instanceof Error ? error.message : String(error);
     console.warn(
-      `Failed to save phone for issuer ${companyName}:`,
-      issuer.issuerPhoneNumber,
-      error
+      `Failed to save phone "${issuer.issuerPhoneNumber}" for issuer ${companyName}: ${message}`
     );
   }
